Remove margins from the Visio diagrams.
Visio removes the margins when exporting to PNG, but when exporting to
SVG, well, things seem to be a bit more... subtle. To avoid
inconsistencies between the two formats (which makes for a jarring user
experience), remove the margins in the Visio diagrams and leave a
README as a cautionary tale.