Remove mention of standing waves in the interference post.
The post claimed that the pattern of constructive and destructive
interference shown in the examples constitutes a standing wave. That's
only half-true; the wave doesn't travel much horizontally, but it still
travels vertically. This is not a standing wave in the 2-D (or 3-D)
sense of the term; it is only a standing wave if one looks at a (mostly)
horizontal 1-D slice of the plot.